update: CMakeLists
I had some trouble with new files actually building, i.e., I would add a new .h/.cpp pair to the project and performing the make command would not build it. If I updated the CMakeLists in any way it would then detect the new cpp and compile it. Don't know enough about CMake to know whats wrong. I think as long as a new source file is added here in the CMakeLists it will compile