update: FileSystemEntry struct declaration/ definition
Every entry in the root directory (including the root directory itself) will be mapped into a file system entry. Importantly each entry has a pointer to a parent directory. This is how I hope to map the tree like strcture of the file system into the linear data structure of an array.