update: recursive method created to read directory structure
this recursive function can be used to read the root directory (or any directory) and any sub-directory it encounters it will read the contents of that directory too, adding any file/ directory it encounters to the array of directory entries. Two helper functions are also added. One is used to subtract two 4 byte (32 bit) addresses, similar to the addition helper function I created. The other helper function is used to convert a cluster number to a sector address (that can be passed to CMD17) to be read from sd memory