update: cmd24 data response token conditional check
A check for 0xCA was done originally because an sd card was having an off by 1 bit error, this was a quick fix to actually make the function return normally. Attempts were made to get this sd card to work but none worked, retesting this cmd with a higher quality sandisk sd card showed that the implementation was in fact correct. Therefore this quick 'fix' was removed